feat: implemnts usecase to scheduler a stats

+++ internal/models/stat.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Stat struct {
-	Users     int64     `json:"users"`
-	Category  int64     `json:"category"`
-	Article   int64     `json:"article"`
+	Users     int       `json:"users"`
+	Category  int       `json:"category"`
+	Article   int       `json:"article"`
 	CreatedAt time.Time `json:"created_at"`
 }
